MT6580: FM Radio code
The code i dropped in some time ago just appeared to WORK!!

The remaining problems were mostly related to the audio subsystem...

In short, the FM radio is an I2S master while the AFE needs to be
configured as the I2S slave.

But i had it set to master mode, and yes it did work but after
some time it will desync and the sound will transition into noise
after some time (i.e. it will slowly shift the "mostly noise" LSBs
into the MSBs) -- and then it will transition back to normal.

But setting the AFE's I2S into slave mode didn't produce anything at
all, i thought it was an issue with the CONNSYS itself, but it actually
turned out that the ASRC "Audio Sample Rate Conversion" was needed to
be enabled, and now, everything is working great!!